feat: RocketResponse Json Successfully Retrieved
- Bug Fix: Certain ints needed to be doubles within RocketResponse models

+++ app/src/main/java/com/example/tin/spacexrockets/models/Diameter.java
@@ -12,10 +12,10 @@ public class Diameter implements Parcelable
 
     @SerializedName("meters")
     @Expose
-    private int meters;
+    private double meters;
     @SerializedName("feet")
     @Expose
-    private int feet;
+    private double feet;
     public final static Creator<Diameter> CREATOR = new Creator<Diameter>() {
 
 
@@ -34,26 +34,26 @@ public class Diameter implements Parcelable
     ;
 
     protected Diameter(Parcel in) {
-        this.meters = ((int) in.readValue((int.class.getClassLoader())));
-        this.feet = ((int) in.readValue((int.class.getClassLoader())));
+        this.meters = ((double) in.readValue((double.class.getClassLoader())));
+        this.feet = ((double) in.readValue((double.class.getClassLoader())));
     }
 
     public Diameter() {
     }
 
-    public int getMeters() {
+    public double getMeters() {
         return meters;
     }
 
-    public void setMeters(int meters) {
+    public void setMeters(double meters) {
         this.meters = meters;
     }
 
-    public int getFeet() {
+    public double getFeet() {
         return feet;
     }
 
-    public void setFeet(int feet) {
+    public void setFeet(double feet) {
         this.feet = feet;
     }
